Compile list of classes to restrict with linter
The `DisallowViewComponentHtmlClasses` linter will scan ERB files for
use of classes that we want to reserve for internal use by Primer
ViewComponents. This commit augments the `export-css-selectors` script
to also generate a list of such classes that we want to restrict. Those
classes are written out to `static/classes.json` and read into a
constant when the linter is loaded.

The script will also add classes to the individual `*.css.json` files
that live alongside our CSS files for test purposes. Most of these json
files are not committed.